1.0.3.pre1
This version contains experimental Gear Score system. A score is calculated based on players level, it's attribute value and the main attribute type. Penalization is applied for items that do not match wearers primary attribute. Gems are not counted in yet, but will be in the future.
Resulting scores are averaged, and should be in range of 0 - 100 (100+ can appear if player has better items than his level - since item attributes are set by level with range added to it).

- Removed unnecessary step when importing HAR files
- Fixed bug in python script